+++ lib/pretest/config/mac.rb
@@ -17,12 +17,16 @@ module MacOS
   end
 
   def mac_gecko_installer
-    puts 'Installing GeckoDriver'
-    system 'curl -OL https://github.com/mozilla/geckodriver/releases/download/v0.15.0/geckodriver-v0.15.0-macos.tar.gz'
-    system 'sudo tar xvjf geckodriver-v0.15.0-macos.tar.gz > /dev/null'
-    system 'sudo chmod +x geckodriver > /dev/null'
-    system 'sudo mv -f geckodriver /usr/local/share/geckodriver > /dev/null'
-    system 'sudo rm -rf geckodriver-v0.15.0-macos.tar.gz > /dev/null'
+    if Dir.entries('/usr/local/share').to_s.include?('geckodriver')
+      puts 'GeckoDriver is already installed on the current system'
+    else
+      puts 'Installing GeckoDriver'
+      system 'curl -OL https://github.com/mozilla/geckodriver/releases/download/v0.15.0/geckodriver-v0.15.0-macos.tar.gz'
+      system 'sudo tar xvjf geckodriver-v0.15.0-macos.tar.gz > /dev/null'
+      system 'sudo chmod +x geckodriver > /dev/null'
+      system 'sudo mv -f geckodriver /usr/local/share/geckodriver > /dev/null'
+      system 'sudo rm -rf geckodriver-v0.15.0-macos.tar.gz > /dev/null'
+    end
     check_gecko_environment_variables
   end
 end
